requirements.txt
openai>=1.0.0
python-dotenv>=1.0.0
gitpython>=3.1.40
PyGithub>=2.1.1
aiohttp>=3.9.1
tiktoken>=0.5.1
pydantic>=2.5.2
rich>=13.7.0
tomli>=2.0.1
pathspec>=0.11.2
packaging>=23.2